refactor(audio): move encode_wav into audio/wav.rs
It is a pure PCM-to-WAV encoder with no ties to the cpal capture
machinery it lived next to. Callers now use the audio::wav path.
